Improve flexnet cache and reduce duplicate code
- Correct path for cargo build cache with rootless docker
- Move verify ledger and count-tx script to common
- Update README.md with current workflows
- Devnet integration: expose port within docker network when testing to
  avoid port conflicts

+++ monad-scripts/jenkins/flexnet-nightly.sh
@@ -19,7 +19,13 @@ fi
 git submodule update --init common/blst
 
 rm -rf logs && mkdir -p logs
+set +e
 nets/net2/scripts/net-run.sh --output-dir logs --net-dir nets/net2/ --flexnet-root . --monad-bft-root ../.. 
+net2_status=$?
+
+if [[ $net2_status -ne 0 ]]; then
+    exit 1
+fi
 
 # clean up the artifacts if the test succeeeds
 rm -rf logs
